Refactored ClueWeb09{Collection,Document} into single class with inner classes.
